proto: remove vote serializers
